+++ CHANGELOG.md
@@ -6,10 +6,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 ### Added
+ - Announce queue times, silence, and connection changes on sports, sportsremote, and remote broadcasts. [Issue 19](https://github.com/Lovinity/wwsu-dj-controls/issues/19)
 
 ### Changed
 
 ### Fixed
+ - DJs on the air might not get a silence detection warning if the DJ Controls they were using does not have client.emergencies being true.
 
 ## [4.5.1] - 2018-12-17
 ### Added